+++ src/main/groovy/ru/vyarus/gradle/plugin/python/task/PythonTask.groovy
@@ -2,7 +2,6 @@ package ru.vyarus.gradle.plugin.python.task
 
 import groovy.transform.CompileStatic
 import org.gradle.api.GradleException
-import org.gradle.api.logging.LogLevel
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.TaskAction
@@ -23,16 +22,11 @@ import ru.vyarus.gradle.plugin.python.cmd.Python
 class PythonTask extends BasePythonTask {
 
     /**
-     * Working directory. Not required, but could be useful for some modules (e.g. generators).
-     */
-    @Input
-    @Optional
-    String workDir
-    /**
      * Create work directory if it doesn't exist. Enabled by default.
      */
     @Input
     boolean createWorkDir = true
+
     /**
      * Module name. If specified, "-m module " will be prepended to specified command (if command not specified then
      * modules will be called directly).
@@ -40,6 +34,7 @@ class PythonTask extends BasePythonTask {
     @Input
     @Optional
     String module
+
     /**
      * Python command to execute. If module name set then it will be module specific command.
      * Examples:
@@ -53,41 +48,27 @@ class PythonTask extends BasePythonTask {
     @Input
     @Optional
     Object command
-    /**
-     * Python logs output level. By default it's {@link LogLevel@LIFECYCLE} (visible with '-i' gradle flag).
-     */
-    @Input
-    @Optional
-    LogLevel logLevel = LogLevel.LIFECYCLE
-    /**
-     * Python arguments applied to all executed commands. Arguments applied before called command
-     * (and so option may be useful for cases impossible with {@link #extraArgs}, applied after command).
-     * For example, it could be used for -I or -S flags (be aware that -S can cause side effects, especially
-     * inside virtual environments).
-     */
-    @Input
-    @Optional
-    List<String> pythonArgs = []
-    /**
-     * Extra arguments to append to every called command.
-     * Useful for pre-configured options, applied to all executed commands
-     */
-    @Input
-    @Optional
-    List<String> extraArgs = []
-    /**
-     * Environment variables for executed python process (variables specified in gradle's
-     * {@link org.gradle.process.ExecSpec#environment(java.util.Map)} during python process execution).
-     */
-    @Input
-    @Optional
-    Map<String, Object> environment = [:]
+
     /**
      * Prefix each line of python output. By default it's '\t' to indicate command output.
      */
     @Input
     @Optional
     String outputPrefix = '\t'
+
+    /**
+     * Extra arguments to append to every called command.
+     * Useful for pre-configured options, applied to all executed commands
+     * <p>
+     * Option not available in {@link BasePythonTask} because of pip tasks which use different set of keys
+     * for various commands. Special pip tasks like {@link ru.vyarus.gradle.plugin.python.task.pip.PipInstallTask}
+     * use multiple different calls internally and general extra args would apply to all of them and, most likely,
+     * crash the build. It is better to implement external arguments support on exact task level (to properly apply it
+     * to exact executed command and avoid usage confusion).
+     */
+    @Input
+    @Optional
+    List<String> extraArgs = []
 
     @TaskAction
     void run() {
@@ -99,29 +80,13 @@ class PythonTask extends BasePythonTask {
         initWorkDirIfRequired()
 
         Python python = python
-                .logLevel(getLogLevel())
                 .outputPrefix(getOutputPrefix())
-                .workDir(getWorkDir())
-                .pythonArgs(getPythonArgs())
                 .extraArgs(getExtraArgs())
-                .environment(environment)
 
         if (mod) {
             python.callModule(mod, cmd)
         } else {
             python.exec(cmd)
-        }
-    }
-
-    /**
-     * Add python arguments, applied before command.
-     *
-     * @param args arguments
-     */
-    @SuppressWarnings('ConfusingMethodName')
-    void pythonArgs(String... args) {
-        if (args) {
-            getPythonArgs().addAll(args)
         }
     }
 
@@ -137,30 +102,7 @@ class PythonTask extends BasePythonTask {
         }
     }
 
-    /**
-     * Add environment variable for python process (will override previously set value).
-     *
-     * @param var variable name
-     * @param value variable value
-     */
-    @SuppressWarnings('ConfusingMethodName')
-    void environment(String var, Object value) {
-        getEnvironment().put(var, value)
-    }
-
-    /**
-     * Add environment variables for python process (will override already set values, but not replace context
-     * map completely). May be called multiple times: all variables would be aggregated.
-     *
-     * @param vars (may be null)
-     */
-    @SuppressWarnings('ConfusingMethodName')
-    void environment(Map<String, Object> vars) {
-        if (vars) {
-            getEnvironment().putAll(vars)
-        }
-    }
-
+    @SuppressWarnings('UnnecessaryGetter')
     private void initWorkDirIfRequired() {
         String dir = getWorkDir()
         if (dir && isCreateWorkDir()) {