+++ src/main/groovy/ru/vyarus/gradle/plugin/python/task/pip/BasePipTask.groovy
@@ -111,7 +111,7 @@ class BasePipTask extends BasePythonTask {
     @Memoized
     @SuppressWarnings('UnnecessaryGetter')
     protected Pip getPip() {
-        Pip pip = new Pip(project, getPythonPath(), getPythonBinary(), getUserScope(), getUseCache())
+        Pip pip = new Pip(python, getUserScope(), getUseCache())
         pip.trustedHosts = getTrustedHosts()
         pip.extraIndexUrls = getExtraIndexUrls()
         pip